moved polciy to own model to allow it getting persisted properly
