+++ examples/restaurantbot/bot.py
@@ -7,19 +7,19 @@ import argparse
 import logging
 import warnings
 
+from policy import RestaurantPolicy
 from rasa_core import utils
 from rasa_core.actions import Action
 from rasa_core.agent import Agent
 from rasa_core.channels.console import ConsoleInputChannel
-from rasa_core.interpreter import RasaNLUInterpreter
-from rasa_core.policies.keras_policy import KerasPolicy
-from rasa_core.policies.memoization import MemoizationPolicy
 from rasa_core.events import SlotSet
+from rasa_core.interpreter import RasaNLUInterpreter
+from rasa_core.policies.memoization import MemoizationPolicy
 
 logger = logging.getLogger(__name__)
 
 
-class RestaurantAPI:
+class RestaurantAPI(object):
     def search(self, info):
         return "papi's pizza place"
 
@@ -46,30 +46,6 @@ class ActionSuggest(Action):
                                  "hint: I'm not going to "
                                  "find anything else :)")
         return []
-
-
-class RestaurantPolicy(KerasPolicy):
-    def model_architecture(self, num_features, num_actions, max_history_len):
-        """Build a Keras model and return a compiled model."""
-        from keras.layers import LSTM, Activation, Masking, Dense
-        from keras.models import Sequential
-
-        n_hidden = 32  # size of hidden layer in LSTM
-        # Build Model
-        batch_shape = (None, max_history_len, num_features)
-
-        model = Sequential()
-        model.add(Masking(-1, batch_input_shape=batch_shape))
-        model.add(LSTM(n_hidden, batch_input_shape=batch_shape))
-        model.add(Dense(input_dim=n_hidden, output_dim=num_actions))
-        model.add(Activation('softmax'))
-
-        model.compile(loss='categorical_crossentropy',
-                      optimizer='adam',
-                      metrics=['accuracy'])
-
-        logger.debug(model.summary())
-        return model
 
 
 def train_dialogue(domain_file="restaurant_domain.yml",
